% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar.R
\name{check_login}
\alias{check_login}
\title{Verifica se o login no site do TJSP está ativo}
\usage{
check_login(cookies)
}
\arguments{
\item{cookies}{Um objeto de cookies (geralmente do tipo `httr2::cookie`) que representa
os dados de sessão autenticada previamente salvos.}
}
\value{
Um valor lógico (`TRUE` ou `FALSE`). Retorna `TRUE` se o login estiver válido (sessão ativa)
e `FALSE` caso contrário.
}
\description{
Esta função realiza uma requisição ao endpoint de verificação de login do site
do Tribunal de Justiça de São Paulo (TJSP) e verifica se a sessão associada aos
cookies fornecidos está autenticada.
}
