% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar.R
\name{cookies}
\alias{cookies}
\title{Obter caminho do arquivo de cookies para uso em requisições}
\usage{
cookies(cookies_path = NULL)
}
\arguments{
\item{cookies_path}{Caminho para o arquivo de cookies. Se `NULL`, tenta recuperar o cookie
da última requisição HTTP feita pela sessão do `httr2`.}
}
\value{
Um `character` com o caminho do arquivo de cookies.
}
\description{
Essa função retorna o caminho do arquivo de cookies a ser utilizado em requisições,
seja pelo caminho fornecido pelo usuário ou, se não fornecido, pelo último cookie
usado na sessão atual do pacote.
}
