% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eproc_lista.R
\name{eproc_lista_baixar}
\alias{eproc_lista_baixar}
\title{Baixa lista de processos distribuídos do sistema e-proc TJSP}
\usage{
eproc_lista_baixar(
  dt_ini = NULL,
  dt_fim = NULL,
  diretorio = ".",
  cookies_path = NULL
)
}
\arguments{
\item{dt_ini}{Data inicial do período de consulta no formato "dd/mm/aaaa".}

\item{dt_fim}{Data final do período de consulta no formato "dd/mm/aaaa".}

\item{diretorio}{Caminho do diretório onde o arquivo será salvo.
Padrão é "." (diretório atual).}

\item{cookies_path}{Caminho para o arquivo de cookies. Se NULL, será criado
um arquivo temporário.}
}
\value{
Não retorna valor. Salva arquivo HTML no diretório especificado com
  nome no formato "ini\{data\}_fim\{data\}.html".
}
\description{
Esta função realiza o download da lista de processos distribuídos em um
período específico do sistema e-proc do Tribunal de Justiça de São Paulo.
O arquivo HTML resultante é salvo no diretório especificado.
}
