% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cpopg_baixar.R
\name{esaj_cpopg_baixar}
\alias{esaj_cpopg_baixar}
\title{Baixar dados da consulta processual de primeiro grau}
\usage{
esaj_cpopg_baixar(processos = NULL, diretorio = ".", cookies_path = NULL)
}
\arguments{
\item{processos}{Númeração Única do processo (CNJ) ou Código do Processo}

\item{diretorio}{Diretório onde será armazenado o html}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau
}
\examples{
\dontrun{
Exemplo típico de processo
esaj_cpopg_baixar("10031013820238260223")

Exemplo de processo em que um único processo gera uma lista de processos
esaj_cpopg_baixar("00002359320188260047")

Exemplo típico de cd_processo
esaj_cpopg_baixar("JK0001UPV0000")

Exemplos de cd_processo de incidente
esaj_cpopg_baixar("JK00024VK0000")
esaj_cpopg_baixar(c("10091165420168260001", "WW0015T010000"))
}

}
