% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cpopg_baixar.R
\name{esaj_cpopg_baixar_cd_processo}
\alias{esaj_cpopg_baixar_cd_processo}
\title{Baixar dados da consulta processual de primeiro grau com base no código do processo}
\usage{
esaj_cpopg_baixar_cd_processo(
  cd_processo,
  diretorio = ".",
  cookies_path = NULL
)
}
\arguments{
\item{cd_processo}{Código do processo}

\item{diretorio}{Diretório onde armazenar os arquivos.}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau com base no código do processo
}
\examples{
\dontrun{
Exemplo típico
esaj_cpopg_baixar_cd_processo("67000FE100000")

Exemplo de incidente
esaj_cpopg_baixar_cd_processo("1B00023PT1PQ8")
}

}
