% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cpopg_outros.R
\name{esaj_cpopg_baixar_outros}
\alias{esaj_cpopg_baixar_outros}
\title{Iteração da função esaj_cpopg_baixar_outros_unitario()}
\usage{
esaj_cpopg_baixar_outros(
  consultas = NULL,
  parametro = c("NMPARTE", "DOCPARTE", "NMADVOGADO", "NUMOAB", "PRECATORIA", "DOCDELEG",
    "NUMCDA"),
  diretorio = ".",
  cookies_path = NULL
)
}
\arguments{
\item{consultas}{Vetor de valores a serem consultados}

\item{parametro}{Parâmetro a ser utilizado na consulta.
Se nenhum parâmetro for informado, o default é considerar 'NUMPROC'.
'NMPARTE' = Nome da parte;
'DOCPARTE' = Número do documento da parte;
'NMADVOGADO' = Nome do advogado;
'NUMOAB' = Número da OAB do advogado;
'PRECATORIA' = Número do precatório;
'DOCELEG' = Número do documento de delegacia;
'NUMCDA' = Número de Certidão de Dívida Ativa (CDA)}

\item{diretorio}{Diretório onde serão armazenados os htmls}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Iteração da função esaj_cpopg_baixar_outros_unitario()
}
\details{
A estrutura do nome dos arquivos gerados é sempre
./\{parametro\}_\{consulta\}_\{pagina\}.html.
}
\examples{
\dontrun{
esaj_cpopg_baixar_outros(consultas = c("123456SP", "100000SP"), parametro = "NUMOAB")
esaj_cpopg_baixar_outros(consultas = "José da Silva Pereira", parametro = "NMPARTE")
}

}
