% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cpopg_baixar.R
\name{esaj_cpopg_baixar_processo}
\alias{esaj_cpopg_baixar_processo}
\title{Baixar dados da consulta processual de primeiro grau com base na numeração única do processo}
\usage{
esaj_cpopg_baixar_processo(
  processo = NULL,
  diretorio = ".",
  cookies_path = NULL
)
}
\arguments{
\item{processo}{Númeração Única do processo (CNJ)}

\item{diretorio}{Diretório onde será armazenado o html}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau com base na numeração única do processo
}
\examples{
\dontrun{
Exemplo típico
esaj_cpopg_baixar_processo("10031013820238260223")

Exemplo em que um único processo gera uma lista de processos
esaj_cpopg_baixar_processo("00002359320188260047")
}

}
