% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cpopg_outros.R
\name{esaj_cpopg_ler_outros}
\alias{esaj_cpopg_ler_outros}
\title{Iteração da função esaj_cpopg_ler_outros_unitario()}
\usage{
esaj_cpopg_ler_outros(arquivos)
}
\arguments{
\item{arquivos}{Vetor contendo o path de arquivos html}
}
\value{
tibble contendo 10 colunas
}
\description{
Iteração da função esaj_cpopg_ler_outros_unitario()
}
