% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_baixar.R
\name{esaj_cposg_baixar}
\alias{esaj_cposg_baixar}
\title{Baixar dados da consulta processual de segundo grau}
\usage{
esaj_cposg_baixar(processos = NULL, diretorio = ".", cookies_path = NULL)
}
\arguments{
\item{processos}{Númeração Única do processo (CNJ) ou Código do Processo}

\item{diretorio}{Diretório onde será armazenado o html}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de segundo grau
}
\examples{
\dontrun{
Exemplo de processo de segundo grau
esaj_cposg_baixar("20598403820228260000")

Exemplo de processo de primeiro grau que gera lista de processos de segundo grau
esaj_cposg_baixar("10001414820158260625")

Exemplo de cd_processo
esaj_cposg_baixar("RI004CVB30000")
}

}
