% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_baixar.R
\name{esaj_cposg_baixar_cd_processo}
\alias{esaj_cposg_baixar_cd_processo}
\title{Baixar dados da consulta processual de segundo grau com base no código do processo}
\usage{
esaj_cposg_baixar_cd_processo(
  cd_processo = NULL,
  diretorio = ".",
  cookies_path = "."
)
}
\arguments{
\item{cd_processo}{Código do processo}

\item{diretorio}{Diretório onde armazenar os arquivos.}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de segundo grau com base no código do processo
}
\examples{
\dontrun{
Exemplo de cd_processo
esaj_cposg_baixar("RI004CVB30000")
}

}
