% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_baixar.R
\name{esaj_cposg_baixar_processo}
\alias{esaj_cposg_baixar_processo}
\title{Baixar dados da consulta processual de segundo grau com base na numeração única do processo}
\usage{
esaj_cposg_baixar_processo(
  processo = NULL,
  diretorio = ".",
  cookies_path = NULL
)
}
\arguments{
\item{processo}{Númeração Única do processo (CNJ)}

\item{diretorio}{Diretório onde será armazenado o html}

\item{cookies_path}{Caminho para salver os cookies (usar somente em casos de paralização)}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de segundo grau com base na numeração única do processo
}
\examples{
\dontrun{
Exemplo de processo de segundo grau
esaj_cposg_baixar("20598403820228260000")

Exemplo de processo de primeiro grau que gera lista de processos de segundo grau
esaj_cposg_baixar("10001414820158260625")
}

}
