% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_ler.R
\name{esaj_cposg_ler_apensos}
\alias{esaj_cposg_ler_apensos}
\title{Ler Apensos no CPOSG}
\usage{
esaj_cposg_ler_apensos(arquivo = NULL)
}
\arguments{
\item{arquivo}{Caminho para o arquivo HTML ou objeto compatível com `xml2::read_html()`.}
}
\value{
Um tibble com as seguintes colunas:
\describe{
   \item{processo}{Número do processo, somente dígitos.}
   \item{cd_processo}{Código interno do processo.}
   \item{preencher}{A preencher.}
}
}
\description{
Extrai tabela de apensos do CPOSG.
}
