% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_ler.R
\name{esaj_cposg_ler_capa}
\alias{esaj_cposg_ler_capa}
\title{Ler dados da capa do processo no CPOSG}
\usage{
esaj_cposg_ler_capa(arquivo = NULL)
}
\arguments{
\item{arquivo}{Caminho para o arquivo HTML ou objeto compatível com `xml2::read_html()`.}
}
\value{
Um tibble com as seguintes colunas:
\describe{
   \item{processo}{Número do processo, somente dígitos.}
   \item{cd_processo}{Código interno do processo.}
   \item{situacao}{Situação atual do processo.}
   \item{classe}{Classe do processo.}
   \item{assunto}{Assunto principal do processo.}
   \item{secao}{Seção judicial responsável.}
   \item{orgao_julgador}{Órgão julgador do processo.}
   \item{area}{Área temática do processo.}
   \item{relator}{Nome do relator do processo.}
   \item{valor_da_acao}{Valor da ação no processo.}
}
}
\description{
Extrai informações básicas da capa do processo judicial a partir do arquivo HTML.
}
