% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_ler.R
\name{esaj_cposg_ler_numeros_pg}
\alias{esaj_cposg_ler_numeros_pg}
\title{Ler tabela de números de primeira instância no CPOSG}
\usage{
esaj_cposg_ler_numeros_pg(arquivo = NULL)
}
\arguments{
\item{arquivo}{Caminho para o arquivo HTML ou objeto compatível com `xml2::read_html()`.}
}
\value{
Um tibble com as colunas:
\describe{
   \item{processo}{Número do processo principal, somente dígitos.}
   \item{cd_processo}{Código interno do processo principal.}
   \item{processo_origem}{Número do processo de primeira instância, somente dígitos.}
   \item{foro_origem}{Foro de origem do processo.}
   \item{vara_origem}{Vara de origem do processo.}
   \item{juiz_origem}{Nome do juiz responsável.}
   \item{obs}{Observações relacionadas.}
}
}
\description{
Extrai dados dos processos de primeira instância vinculados ao processo principal a partir do arquivo HTML.
}
