% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_ler.R
\name{esaj_cposg_ler_partes}
\alias{esaj_cposg_ler_partes}
\title{Ler partes do processo no CPOSG}
\usage{
esaj_cposg_ler_partes(arquivo = NULL)
}
\arguments{
\item{arquivo}{Caminho para o arquivo HTML ou objeto compatível com `xml2::read_html()`.}
}
\value{
Um tibble com as colunas:
\describe{
   \item{processo}{Número do processo, somente dígitos.}
   \item{cd_processo}{Código interno do processo.}
   \item{da_pg}{Tibble com dados das partes contendo:}
   \item{id_parte}{Identificador único para cada parte.}
   \item{participacao}{Tipo de participação da parte (ex: autor, réu).}
   \item{nome}{Nome da parte ou do advogado.}
   \item{tipo_parte}{Tipo da parte (ex: Principal, Advogado).}
}
}
\description{
Extrai as partes relacionadas ao processo, incluindo advogados e tipos de participação.
}
