% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaj_cposg_ler.R
\name{esaj_cposg_ler_subprocessos_recursos}
\alias{esaj_cposg_ler_subprocessos_recursos}
\title{Ler subprocessos e recursos do processo no CPOSG}
\usage{
esaj_cposg_ler_subprocessos_recursos(arquivo = NULL)
}
\arguments{
\item{arquivo}{Caminho para o arquivo HTML ou objeto compatível com `xml2::read_html()`.}
}
\value{
Um tibble com as seguintes colunas:
\describe{
   \item{processo}{Número do processo, somente dígitos.}
   \item{cd_processo}{Código interno do processo.}
   \item{dt_recebimento}{Data de recebimento do recurso.}
   \item{classe}{Classe do recurso.}
   \item{url}{URL no CPOSG do recurso}
}
}
\description{
Verifica a existência de subprocessos e recursos associados ao processo.
}
