% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autenticar.R
\name{get_email_token}
\alias{get_email_token}
\title{Obtém tokens do email}
\usage{
get_email_token(email_provider, outlook, hora)
}
\arguments{
\item{email_provider}{"gmail" ou "outlook"}

\item{outlook}{Se outlook, informar se "business" ou
"personal"}

\item{hora}{Parâmetro não usado, por ora.}
}
\value{
token
}
\description{
Obtém tokens do email
}
