% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_baixar.R
\name{tjsp_cpopg_baixar}
\alias{tjsp_cpopg_baixar}
\title{Baixar dados da consulta processual de primeiro grau}
\usage{
tjsp_cpopg_baixar(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Númeração Única do processo (CNJ) ou Código do Processo}

\item{diretorio}{Diretório onde será armazenado o html}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau
}
\examples{
\dontrun{
tjsp_cpopg_baixar("10031013820238260223") # exemplo típico de processo
tjsp_cpopg_baixar("00002359320188260047") # exemplo de processo em que um único processo gera uma lista de processos
tjsp_cpopg_baixar("JK0001UPV0000") # exemplo típico de cd_processo
tjsp_cpopg_baixar("JK00024VK0000") # exemplo de cd_processo de incidente
tjsp_cpopg_baixar(c("10091165420168260001", "WW0015T010000"))
}

}
