% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_baixar.R
\name{tjsp_cpopg_baixar_cd_processo}
\alias{tjsp_cpopg_baixar_cd_processo}
\title{Baixar dados da consulta processual de primeiro grau com base no código do processo}
\usage{
tjsp_cpopg_baixar_cd_processo(cd_processo, diretorio = ".")
}
\arguments{
\item{cd_processo}{Código do processo}

\item{diretorio}{Diretório onde armazenar os arquivos.}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau com base no código do processo
}
\examples{
\dontrun{
tjsp_cpopg_baixar_cd_processo("67000FE100000") # exemplo típico
tjsp_cpopg_baixar_cd_processo("1B00023PT1PQ8") # exemplo de incidente
}

}
