% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_baixar_outros.R
\name{tjsp_cpopg_baixar_outros}
\alias{tjsp_cpopg_baixar_outros}
\title{Iteração da função tjsp_cpopg_baixar_outros_unitario()}
\usage{
tjsp_cpopg_baixar_outros(
  consultas = NULL,
  parametro = c("NMPARTE", "DOCPARTE", "NMADVOGADO", "NUMOAB", "PRECATORIA", "DOCDELEG",
    "NUMCDA"),
  diretorio = "."
)
}
\arguments{
\item{consultas}{Vetor de valores a serem consultados}

\item{parametro}{Parâmetro a ser utilizado na consulta.
Se nenhum parâmetro for informado, o default é considerar 'NUMPROC'.
'NMPARTE' = Nome da parte;
'DOCPARTE' = Número do documento da parte;
'NMADVOGADO' = Nome do advogado;
'NUMOAB' = Número da OAB do advogado;
'PRECATORIA' = Número do precatório;
'DOCELEG' = Número do documento de delegacia;
'NUMCDA' = Número de Certidão de Dívida Ativa (CDA)}

\item{diretorio}{Diretório onde serão armazenados os htmls}
}
\value{
html da consulta
}
\description{
Iteração da função tjsp_cpopg_baixar_outros_unitario()
}
\examples{
\dontrun{
tjsp_cpopg_baixar_outros(consultas = c("123456SP", "100000SP"), parametro = "NUMOAB")
tjsp_cpopg_baixar_outros(consultas = "José da Silva Pereira", parametro = "NMPARTE")
}
}
