% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_baixar_outros.R
\name{tjsp_cpopg_baixar_outros_unitario}
\alias{tjsp_cpopg_baixar_outros_unitario}
\title{Baixar dados da consulta processual de primeiro grau com base em outros parâmetros especificados}
\usage{
tjsp_cpopg_baixar_outros_unitario(
  consulta = NULL,
  parametro = c("NMPARTE", "DOCPARTE", "NMADVOGADO", "NUMOAB", "PRECATORIA", "DOCDELEG",
    "NUMCDA"),
  diretorio = "."
)
}
\arguments{
\item{consulta}{Valor da consulta}

\item{parametro}{Parâmetro a ser utilizado na consulta.
Se nenhum parâmetro for informado, o default é considerar 'NUMPROC'.
'NMPARTE' = Nome da parte;
'DOCPARTE' = Número do documento da parte;
'NMADVOGADO' = Nome do advogado;
'NUMOAB' = Número da OAB do advogado;
'PRECATORIA' = Número do precatório;
'DOCELEG' = Número do documento de delegacia;
'NUMCDA' = Número de Certidão de Dívida Ativa (CDA)}

\item{diretorio}{Diretório onde será armazenado o html}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau com base em outros parâmetros especificados
}
\details{
A estrutura do nome dos arquivos gerados é sempre ./{parametro}_{consulta}_{pagina}.html.
}
\examples{
\dontrun{
tjsp_cpopg_baixar_outros_unitario(consulta="123456SP", parametro = "NUMOAB")
tjsp_cpopg_baixar_outros_unitario(consulta = "José da Silva Pereira", parametro = "NMPARTE")
}


}
