% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_baixar.R
\name{tjsp_cpopg_baixar_processo}
\alias{tjsp_cpopg_baixar_processo}
\title{Baixar dados da consulta processual de primeiro grau com base na numeração única do processo}
\usage{
tjsp_cpopg_baixar_processo(processo, diretorio = ".")
}
\arguments{
\item{processo}{Númeração Única do processo (CNJ)}

\item{diretorio}{Diretório onde será armazenado o html}
}
\value{
html da consulta
}
\description{
Baixar dados da consulta processual de primeiro grau com base na numeração única do processo
}
\examples{
\dontrun{
tjsp_cpopg_baixar_processo("10031013820238260223") # exemplo típico
tjsp_cpopg_baixar_processo("00002359320188260047") # exemplo em que um único processo gera uma lista de processos
}

}
