% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_ler.R
\name{tjsp_cpopg_ler}
\alias{tjsp_cpopg_ler}
\title{tjsp_cpopg_ler() parseia os dados de um html do cpopg.}
\usage{
tjsp_cpopg_ler(
  arquivos = NULL,
  formato = c("Padronizado", "Completo"),
  outros = c("Tudo", "Delegacia", "Partes", "Movimentacoes", "Peticoes Diversas",
    "Incidentes", "Apensos", "Audiencias", "Historico Classes")
)
}
\arguments{
\item{arquivos}{Vetor contendo o caminho para os arquivos em html do cpopg.
Pode receber também apenas a pasta em que estão os arquivos.}

\item{outros}{Informações a serem lidas do html. Por default, lê-se tudo ("Tudo").
Se quiser ler apenas os metadados principais, info deve ser NULL.}

\item{capa}{Formato das colunas a serem retornadas. É escolhido 'Padronizado' por default.
* 'Padronizado' retorna uma tibble com 14 colunas, forçando algumas a
aparecerem quando um processo não tiver e excluindo outras colunas
que possam aparecer em um processo, mas não em outros
* 'Completo' retorna uma tibble sem pré-definir as colunas, podendo trazer
resultados distintos para cada processo.}
}
\value{
tibble
}
\description{
tjsp_cpopg_ler() parseia os dados de um html do cpopg.
}
