% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_ler_outros.R
\name{tjsp_cpopg_ler_outros}
\alias{tjsp_cpopg_ler_outros}
\title{Iteração da função tjsp_cpopg_ler_outros_unitario()}
\usage{
tjsp_cpopg_ler_outros(arquivos)
}
\arguments{
\item{arquivos}{Vetor contendo o path de arquivos html}
}
\value{
tibble contendo 10 colunas
}
\description{
Iteração da função tjsp_cpopg_ler_outros_unitario()
}
