% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjsp_cpopg_ler_outros.R
\name{tjsp_cpopg_ler_outros_unitario}
\alias{tjsp_cpopg_ler_outros_unitario}
\title{Lê busca por paramêtro diferente do processo}
\usage{
tjsp_cpopg_ler_outros_unitario(arquivo)
}
\arguments{
\item{arquivo}{Path de um arquivo html}
}
\value{
tibble contendo 10 colunas
}
\description{
Lê busca por paramêtro diferente do processo
}
